+++ composeApp/src/androidMain/kotlin/net/thevenot/comwatt/MainActivity.android.kt
@@ -3,6 +3,7 @@ package net.thevenot.comwatt
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
@@ -12,6 +13,8 @@ import net.thevenot.comwatt.utils.ScreenOrientationController
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        enableEdgeToEdge()
 
         ScreenOrientationController.setActivity(this)